Move code to another file and add some tests
