+++ app/core/RPCMethods/index.js
@@ -1,0 +1,7 @@
+import wallet_addEthereumChain from './wallet_addEthereumChain.js';
+
+const RPCMethods = {
+	wallet_addEthereumChain
+};
+
+export default RPCMethods;
